Archiving error logs in CI build, moved global variables to multiple scripts scope because global variables list is shared among the scripts compiled together
