+++ publish.sh
@@ -1,2 +1,2 @@
 rm -rf ./deploy.zip
-zip -r deploy.zip ./Build/Artifacts/ ./Build/Transpiled/
+zip -r deploy.zip ./Build/Artifacts/ ./Build/Transpiled/ ./Build/compile_log ./Build/error_log
